Move all settings definitions to lib
