+++ html_views/src/goals/goals.ts
@@ -2,13 +2,12 @@ import { WebviewApi } from "vscode-webview";
 import {
   ControllerEvent,
   ResizeEvent,
-  SettingsState,
   ProofViewProtocol,
-  ProofViewDiffSettings,
 } from "./protocol";
 import { Infoview } from "./display-proof-state";
 import "@vscode/webview-ui-toolkit/dist/toolkit";
 import h = require("hyperscript");
+import { ProofViewDiffSettings, SettingsState } from "@lib/settings";
 
 const infoview = Infoview();
 const root = document.getElementById("root");
