+++ html_views/src/goals/protocol.ts
@@ -1,3 +1,5 @@
+import { SettingsState } from "@lib";
+
 export interface ControllerEvent {
   eventName: string;
   params: ResizeEvent // | | | | | ;
@@ -13,20 +15,6 @@ interface GoalUpdate {
 
 interface SettingsUpdate extends SettingsState {
   command: 'settings-update'
-}
-
-export interface ProofViewDiffSettings {
-  enabled: string,
-  addedTextIsItalic : boolean,
-  removedTextIsStrikedthrough : boolean,
-}
-
-export interface SettingsState {
-  fontFamily?: string,
-  fontSize?: string,
-  fontWeight?: string,
-  prettifySymbolsMode?: boolean,
-  proofViewDiff? : ProofViewDiffSettings,
 }
 
 export type ProofViewProtocol = GoalUpdate | SettingsUpdate;
