+++ lib/src/index.ts
@@ -1,1 +1,2 @@
 export * from './protocol';
+export * from './settings';
