+++ server/src/CoqProject.ts
@@ -1,5 +1,4 @@
 import {CoqDocument, DocumentCallbacks, TextDocumentItem} from './document';
-import {Settings} from './protocol';
 import * as vscode from 'vscode-languageserver';
 import * as path from 'path';
 import * as fs from 'fs';
@@ -8,7 +7,7 @@ import * as nodeAsync from './util/nodejs-async';
 import {CoqTop} from './coqtop/CoqTop';
 import {CoqTop as CoqTop8} from './coqtop/CoqTop8';
 import stringArgv from 'string-argv';
-import { DocumentSelector } from "@lib";
+import { DocumentSelector, Settings } from "@lib/settings";
 
 
 const coqProjectFileName = '_CoqProject';
