+++ server/src/util/PrettifySymbols.ts
@@ -1,6 +1,7 @@
-import {AnnotatedText, TextAnnotation, Substitution} from '../protocol';
+import {AnnotatedText, TextAnnotation} from '../protocol';
 import {combineAnnotationText, Annotation, mapAnnotation} from './AnnotatedText';
 import * as server from '../server';
+import { Substitution } from '@lib/settings';
